spdxify.py: improve placement of SPDX identifier
Fixes the following issues:
- When multiple copyright lines are separated by a blank line, the SPDX
ID would be added after the first one. It should be added after the
last one (and separated by a blank line in this case).
- 'All rights reserved' should be considered part of the copyright
statement
- If a file contains multiple copyright blocks (that is, copyright
statements, something else, some more copyright statements), the SPDX
ID had better be put at the beginning of the file.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>